Add JSON string Support for MongoDb Business object so queries and Save() operations can be run with string data. Useful for interfacing with external tools (like FoxPro) that can provide JSON, but can't directly access the MongoDb Generics based interface.

+++ Westwind.Utilities/Utilities/DataUtils.cs
@@ -95,8 +95,8 @@ namespace Westwind.Utilities
         /// <returns></returns>
         public static long GenerateUniqueNumericId()
         {
-            byte[] bytes = Guid.NewGuid().ToByteArray();            
-            return BitConverter.ToInt64(bytes, 0);
+            byte[] bytes = Guid.NewGuid().ToByteArray();
+            return (long) BitConverter.ToUInt64(bytes, 0);
         }
 
         static Random rnd = new Random();
